Moved initTexture function to TreeNode to not repeat code
